Moves integration to plugin
