+++ packages/jazz-richtext-prosemirror/src/lib/plugin.ts
@@ -1,6 +1,12 @@
 import { CoRichText } from "jazz-tools";
+import {
+  DOMParser as PMDOMParser,
+  DOMSerializer as PMDOMSerializer,
+  Slice,
+} from "prosemirror-model";
+import { schema } from "prosemirror-schema-basic";
 import { Plugin, PluginKey } from "prosemirror-state";
-// import { type EditorView } from "prosemirror-view";
+import { type EditorView } from "prosemirror-view";
 
 export const jazzPluginKey = new PluginKey("jazz");
 
@@ -12,31 +18,63 @@ export const jazzPluginKey = new PluginKey("jazz");
  * @returns The ProseMirror plugin
  */
 export function createJazzPlugin(text: CoRichText | undefined) {
-  // let view: EditorView | undefined;
+  let view: EditorView | undefined;
 
   return new Plugin({
     key: jazzPluginKey,
 
-    // Initialize the plugin with ProseMirror view
-    // view(editorView) {
-    //   view = editorView;
-    //   return {
-    //     destroy() {
-    //       view = undefined;
-    //     },
-    //   };
-    // },
+    view(editorView) {
+      view = editorView;
+
+      // Subscribe to CoRichText changes
+      if (text) {
+        text.subscribe((newText) => {
+          if (view && newText) {
+            const doc = new DOMParser().parseFromString(
+              newText.toString(),
+              "text/html",
+            );
+            const pmDoc = PMDOMParser.fromSchema(schema).parse(doc);
+            const tr = view.state.tr.replace(
+              0,
+              view.state.doc.content.size,
+              new Slice(pmDoc.content, 0, 0),
+            );
+            tr.setMeta("fromJazz", true);
+            view.dispatch(tr);
+          }
+        });
+      }
+
+      return {
+        destroy() {
+          view = undefined;
+        },
+      };
+    },
 
     state: {
-      // Initialize the plugin with CoRichText
       init() {
         return {
           text,
         };
       },
 
-      // Update CoRichText when ProseMirror document changes
-      apply(_tr, value) {
+      apply(tr, value) {
+        if (tr.getMeta("fromJazz")) {
+          return value;
+        }
+
+        if (tr.docChanged && text) {
+          const doc = tr.doc;
+          const str = new XMLSerializer()
+            .serializeToString(
+              PMDOMSerializer.fromSchema(schema).serializeFragment(doc.content),
+            )
+            .replace(/\sxmlns="[^"]+"/g, "");
+          text.applyDiff(str);
+        }
+
         return value;
       },
     },
